Refactor Upload screen
