+++ Frontend/src/App.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { ToastContainer, toast } from 'react-toastify';
+import { ToastContainer } from 'react-toastify';
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
 import { BrowseScreen } from './Screens/BrowseScreen';
 import { DetailsScreen } from './Screens/DetailsScreen';
 import { NotFound } from './Screens/NotFound';
-import { UploadScreen } from './Screens/UploadScreen';
+import { UploadScreen } from './Screens/UploadScreen/UploadScreen';
 import { UserScreen } from './Screens/UserScreen';
 import "react-toastify/dist/ReactToastify.css";
